Create ChatGPT:: Removing outliers
3. Redoing EDA on Filtered Data

Now, with the outliers removed, you can redo the EDA on df_filtered. Repeat the same analysis steps as before:

Basic Statistics: Recalculate the basic statistics using df_filtered.describe().

Histogram: Plot a histogram of the 'vertex_count' column.
Box Plot: Create a box plot to visualize the spread of the data.
Violin Plot: Make a violin plot to see the distribution of the data.